+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes
 .. ----------
 .. -
 
+- Add click-odoo-copydb
+- Add click-odoo-dropdb
+
 1.2.0 (2018-10-07)
 ------------------
 - Odoo 12 support